feat: Correction de bug sur l'ajout d'un type d'acte; Ajout d'un dialog pour ajouter des types d'actes

+++ lib/model/type_acte.dart
@@ -1,3 +1,5 @@
+import 'package:app_psy/app_psy_utils.dart';
+
 const String tableTypeActe = 'type_acte';
 
 class TypeActeChamps {
@@ -14,7 +16,7 @@ class TypeActeChamps {
 class TypeActe {
   final int? id;
   final String nom;
-  final int prix;
+  final double prix;
 
   const TypeActe({
     this.id,
@@ -25,7 +27,7 @@ class TypeActe {
   TypeActe copy({
     int? id,
     String? nom,
-    int? prix,
+    double? prix,
   }) =>
       TypeActe(
         id: id ?? this.id,
@@ -37,7 +39,7 @@ class TypeActe {
       TypeActe(
         id: json[TypeActeChamps.id] as int?,
         nom: json[TypeActeChamps.nom] as String,
-        prix: json[TypeActeChamps.prix] as int,
+        prix: json[TypeActeChamps.prix] as double,
       );
 
   Map<String, Object?> toJson() =>
